+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fix TODO keyword in comments [Thanks lperson](https://github.com/BeardedBear/bearded-theme/issues/135)
 - Fix const keyword missing for julia [Thanks SabeDoesThings](https://github.com/BeardedBear/bearded-theme/issues/136)
 - Fix quoted style attribute for HTML [Thanks gmurad97](https://github.com/BeardedBear/bearded-theme/issues/139)
+- Fix inserted comment background color in diff editor [Thanks hadirgax](https://github.com/BeardedBear/bearded-theme/issues/144)
 
 ### Enhancements